big refactor to Stepper class
We now have a:

- StepWalker class, responsible for walking a small distance towards a
final destination.

- SpiralNavigator class, which around an area in a spiral. It
determines points to walk to and uses to StepWalker to slowly go from
one point to another.

This architecture enables us to have different Navigators, although it
doesn’t implement a nice Navigator configuration YET.
